Implement gate

+++ src/Commands/CrudController.php
@@ -45,6 +45,7 @@ class CrudController extends Command
 namespace ' . $nameSpace . ';
 
 use Illuminate\Routing\Controller;
+use Illuminate\Support\Facades\Gate;
 use Modules\\' . $moduleName . '\Entities\\' . $crudController . ';
 use Modules\\' . $moduleName . '\http\Repositories' . $classFolder . '\\' . $crudController . 'Repository;
 use Modules\\' . $moduleName . '\Http\Requests\Store' . $crudController . 'Request;
@@ -54,17 +55,22 @@ class ' . $crudController . 'Controller extends Controller
 {
     public function index()
     {
+        abort_if(Gate::denies("' . $lowerCrudController . '-access"), 403);
+
         $' . Str::plural($lowerCrudController) . ' = ' . $crudController . '::paginate(5);
         return view("' . strtolower($moduleName) . '::' . $view . $lowerCrudController . '.index", compact("' . Str::plural($lowerCrudController) . '"));
     }
 
     public function create()
     {
+        abort_if(Gate::denies("' . $lowerCrudController . '-create"), 403);
         return view("' . strtolower($moduleName) . '::' . $view . $lowerCrudController . '.create");
     }
 
     public function store(Store' . $crudController . 'Request $request)
     {
+        abort_if(Gate::denies("' . $lowerCrudController . '-store"), 403);
+
         $' . $lowerCrudController . ' = new ' . $crudController . '();
         $' . $lowerCrudController . ' = ' . $crudController . 'Repository::storeOrUpdate($' . $lowerCrudController . ', $request);
 
@@ -73,16 +79,22 @@ class ' . $crudController . 'Controller extends Controller
 
     public function show(' . $crudController . ' $' . $lowerCrudController . ')
     {
+        abort_if(Gate::denies("' . $lowerCrudController . '-show"), 403);
+
         return view("' . strtolower($moduleName) . '::' . $view . $lowerCrudController . '.show", compact("' . $lowerCrudController . '"));
     }
 
     public function edit(' . $crudController . ' $' . $lowerCrudController . ')
     {
+        abort_if(Gate::denies("' . $lowerCrudController . '-edit"), 403);
+
         return view("' . strtolower($moduleName) . '::' . $view . $lowerCrudController . '.edit", compact("' . $lowerCrudController . '"));
     }
 
     public function update(Update' . $crudController . 'Request $request, ' . $crudController . ' $' . $lowerCrudController . ')
     {
+        abort_if(Gate::denies("' . $lowerCrudController . '-update"), 403);
+
         $' . $lowerCrudController . ' = ' . $crudController . 'Repository::storeOrUpdate($' . $lowerCrudController . ', $request);
 
         return redirect()->route("' . strtolower($moduleName)  . '.' . $view . $lowerCrudController . '.index")->with("success", $' . $lowerCrudController . '->name . " Updated");
@@ -90,6 +102,7 @@ class ' . $crudController . 'Controller extends Controller
 
     public function destroy(' . $crudController . ' $' . $lowerCrudController . ')
     {
+        abort_if(Gate::denies("' . $lowerCrudController . '-delete"), 403);
         $' . $lowerCrudController . '->delete();
 
         return redirect()->route("' . strtolower($moduleName)  . '.' . $view . $lowerCrudController . '.index")->with("success", $' . $lowerCrudController . '->name . " Deleted!");
@@ -138,8 +151,11 @@ class ' . $crudController . 'Controller extends Controller
             'name' => 'Update' . $crudController . 'Request',
             'module' => $moduleName
         ]);
+        $this->call('wailan:permission', [
+            'class' => $this->argument('controller'),
+        ]);
 
-        $this->line('Please wait untill migration finish');
+        $this->info('Please wait untill migration finish');
 
         $this->call('make:migration', [
             'name' => 'create' . Str::plural($crudController) . '_table'
@@ -163,5 +179,12 @@ class ' . $crudController . 'Controller extends Controller
                 'module' => $this->argument('module')
             ]);
         }
+
+        $this->line('========================================');
+        $this->info('CRUD installation completed');
+        $this->info('Dont forget to');
+        $this->info('Setup database migration and run');
+        $this->info('php artisan migrate --seed');
+        $this->line('========================================');
     }
 }
